Add per-book star user ratings
Adds a user rating feature (1–5 stars) stored on the books table.
Ratings can be set interactively on the book detail page and are imported from Goodreads CSV exports during library import.

The Stars component now supports an interactive mode with hover preview and click-to-rate (clicking the current rating clears it).
BookCover gains an optional rating badge overlay showing a star icon and number, displayed in grid view and the library screen. The list view shows the full inline star display instead.

+++ backend/src/goodreads_importer.rs
@@ -22,9 +22,8 @@ pub struct BookRecord {
     pub isbn: String,
     #[serde(rename = "ISBN13")]
     pub isbn13: String,
-    #[allow(dead_code)]
     #[serde(rename = "My Rating")]
-    pub my_rating: Option<u8>,
+    pub my_rating: Option<f64>,
     #[allow(dead_code)]
     pub publisher: String,
     #[allow(dead_code)]
